requirements.txt
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
firebase-admin==6.2.0
stqdm==0.0.5
Pillow==10.0.1
openai==0.28.1
sentry-sdk==1.28.1
